Move nr_fhi_if4p5_dl_slot_send to 7.2; nr_fhi_if5_dl_slot_send and fh_if5_south_out to AW2S
